pyproject.toml
[project]
name = "langchain-start"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "chromadb>=1.3.7",
    "faiss-cpu>=1.13.1",
    "langchain>=0.2.0",
    "langchain-anthropic>=1.3.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.0",
    "langchain-experimental>=0.4.1",
    "langchain-openai>=1.1.3",
    "langgraph>=1.0.5",
    "langgraph-checkpoint-sqlite>=3.0.1",
    "langgraph-cli>=0.4.11",
    "pydantic>=2.12.5",
    "pypdf>=6.4.2",
    "python-dotenv>=1.2.1",
    "rank-bm25>=0.2.2",
    "streamlit>=1.52.1",
    "tiktoken>=0.12.0",
]
